Move action files into root folder
