+++ src/components/ui/DbLink.js
@@ -1,6 +1,6 @@
 import {Trans} from '@lingui/react'
 import {Provider as TooltipProvider, tooltipHOC} from '@xivanalysis/tooltips'
-import {ITEM_ID_OFFSET} from 'data/ACTIONS/ITEMS'
+import {ITEM_ID_OFFSET} from 'data/ACTIONS'
 import {STATUS_ID_OFFSET} from 'data/STATUSES'
 import {useObserver} from 'mobx-react'
 import PropTypes from 'prop-types'
