+++ src/components/ui/Rotation.tsx
@@ -1,8 +1,7 @@
 import cn from 'classnames'
 import {ActionLink, ItemLink} from 'components/ui/DbLink'
 import {getDataBy} from 'data'
-import ACTIONS from 'data/ACTIONS'
-import {ITEM_ID_OFFSET} from 'data/ACTIONS/ITEMS'
+import ACTIONS, {ITEM_ID_OFFSET} from 'data/ACTIONS'
 import {AbilityEvent} from 'fflogs'
 import PropTypes from 'prop-types'
 import React, {Component} from 'react'
